+++ src/Vector3.h
@@ -7,33 +7,33 @@ namespace math_utils{
     class vec3{
         public:
         vec3();
-        vec3(float x, float y, float z);
+        vec3(double x, double y, double z);
         vec3(const vec3& v){x=v.x; y=v.y; z=v.z;};
 
         vec3 add(const vec3& v);
         vec3 sub(const vec3& v);
         
-        float mod();
+        double mod();
         vec3 norm();
-        float distance(const vec3& v);
+        double distance(const vec3& v);
 
-        float dot_product(const vec3& v);
+        double dot_product(const vec3& v);
         vec3 cross_product(const vec3& v);
         
-        vec3 rotateX(float angle);
-        vec3 rotateY(float angle);
-        vec3 rotateZ(float angle);
-        vec3 scale(float kx, float ky, float kz);
+        vec3 rotateX(double angle);
+        vec3 rotateY(double angle);
+        vec3 rotateZ(double angle);
+        vec3 scale(double kx, double ky, double kz);
         
                 
-        float X(){return x;}
-        float Y(){return y;}
-        float Z(){return z;}
+        double X(){return x;}
+        double Y(){return y;}
+        double Z(){return z;}
 
-        std::vector<float> getV();
+        std::vector<double> getV();
 
         private:
-        float x,y,z;
+        double x,y,z;
 
     };
 }
